backend/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Machine Learning & AI
torch==2.1.0
transformers==4.35.0
scikit-learn==1.3.2
numpy==1.24.3
pandas==2.1.3
scipy==1.11.4

# Computer Vision
opencv-python==4.8.1.78
pytesseract==0.3.10
Pillow==10.1.0

# NLP
sentencepiece==0.1.99
tokenizers==0.15.0
nltk==3.8.1

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.12.1

# Caching
redis==5.0.1
hiredis==2.2.3

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
bcrypt==4.1.1

# HTTP & API
httpx==0.25.1
requests==2.31.0
aiohttp==3.9.1

# Data Validation
email-validator==2.1.0
python-dateutil==2.8.2

# Monitoring & Logging
prometheus-client==0.19.0
python-json-logger==2.0.7

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.1
faker==20.1.0

# Utilities
python-dotenv==1.0.0
click==8.1.7
tqdm==4.66.1
pyyaml==6.0.1

# Model Serving
onnx==1.15.0
onnxruntime==1.16.3

# Data Processing
joblib==1.3.2
pyarrow==14.0.1

# API Documentation
fastapi-pagination==0.12.13